Fixes: - increase stepcount again in RK76 - avoid warning: explicit cast from unsigned int to int - Omit debug printf

+++ Particle_Integrator/Code/SourceCode.c
@@ -97,7 +97,7 @@ int main(int argc, char *argv[])
 			if (j)
 			{
 				printf("\nInput argument set: ");
-				for (k = 0; k < strlen(argv[j]); k++)
+				for (k = 0; k < (int)strlen(argv[j]); k++)
 				{
 					printf("%c", argv[j][k]);
 				}
